AutomatonRenderer major refactoring
The AutomatonRenderer class has been refactored into multiple, smaller
classes. In particular, there's now a MainController class that handles
the communication with the machine controllers and forwards method calls
to the underlying AutomatonRenderer implementation. Basically,
AutomatonRenderer's role is now more compatible with its name: it is
only concerned with the user interface, and non-GUI-related
functionality is handled by the MainController.

The State and Edge classes have been refactored as well: there are
simple interfaces called State and Edge that contain only
non-GUI-related attributes, and the classes that were previously known
as State and Edge are now called UIState and UIEdge, respectively. This
change allows non-GUI-related parts of the application to only have
access to non-GUI-related attributes of states and edges, which improves
separation of concerns. The State and Edge interfaces are located in a
new file called types.ts, which also contains other core types and can
be imported through a reference path.

These changes were merged from the structural_refactoring branch.

Squashed commit of the following:

commit 441c58245748d3d78bab45bc30780d467082e8c2
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Thu Oct 5 21:01:08 2017 -0300

    Renamed Edge/State files; minor cleanup

commit 12cdb6c3e6a96374c74db731dc7ed27f9dea4967
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Thu Oct 5 19:52:45 2017 -0300

    Finished the AutomatonRenderer decoupling

commit 530545fc13ff1819c94f81026b2c2ea0c2730f80
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Thu Oct 5 17:03:56 2017 -0300

    AutomatonRenderer decoupling: part 6

commit f18f656d327a55e2f1a3422354a94ebd77a42111
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Thu Oct 5 12:14:08 2017 -0300

    AutomatonRenderer decoupling: part 5

commit 336a80b6a888ffc1b06550d3f7ee936f933625cc
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Thu Oct 5 00:42:50 2017 -0300

    AutomatonRenderer decoupling: part 4

commit b777019d749639f388988a5837f469d42c2441e8
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Wed Oct 4 21:39:13 2017 -0300

    AutomatonRenderer decoupling: part 3

commit 37570d58f454fe21035ea61eade9df83d15452fd
Author: Ghabriel <ghabriel.nunes@gmail.com>
Date:   Wed Oct 4 20:21:09 2017 -0300

    AutomatonRenderer decoupling: part 2

commit 423dbcaaaa12ec8790f584ae881024cf44461425
Author: Ghabriel Nunes <ghabriel.nunes@gmail.com>
Date:   Wed Oct 4 17:33:31 2017 -0300

    Started decoupling the AutomatonRenderer into multiple classes